backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pdfplumber==0.11.4
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2
alembic==1.13.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
apscheduler>=3.10.0
psycopg2-binary>=2.9.0
anthropic>=0.40.0
openpyxl>=3.1.0
